chore: Drop over-urgent help wanted poster from readme
Help is still welcome and the related issue still pinned, but we have an
active contributer with some maintenence priviledges already. Having
watched the issue for a while I don't think the 'throw your name in for
consideration' is really helping because it attracts people who want the
attention but aren't interested in actually jumping in and doing tho
*hard* parts of maintaining a project.

The glamorious bits are mostly done for this project and the long tail
of issue triage and small bug fixes and improvements is long underway.
Anyone is welcome to jump into those things, but advertising it as an
opening is a) turning people off from using the plugin and b) attracting
the wrong kind of applicants.